Actualizacion

+++ js/login.js
@@ -1,22 +1,25 @@
 
-function showAlertSuccess() {
-   document.getElementById("alert-success").classList.add("show");
-}
-
-function showAlertError() {
-   document.getElementById("alert-danger").classList.add("show");
-}
-
-
 function login(){
     let user = document.getElementById("user").value;
     let pass = document.getElementById("pass").value;
  if(user==="" || pass===""){
     document.getElementById("user").classList.add("error");
     document.getElementById("pass").classList.add("error");
-    showAlertError();
+    Swal.fire({
+      position: 'center',
+      icon: 'error',
+      title: 'Debe rellenar los campos',
+      showConfirmButton: false,
+      timer: 1500
+    })
  } else {
-    showAlertSuccess();
+   Swal.fire({
+      position: 'center',
+      icon: 'success',
+      title: 'Has iniciado sesión',
+      showConfirmButton: false,
+      timer: 1500
+    })
     sessionStorage.setItem("user", user);
     location.href="index.html";
  }   
@@ -27,5 +30,6 @@ document.addEventListener("DOMContentLoaded",()=>{
       let usuario = localStorage.getItem(user);
     document.getElementById("user").innerHTML=usuario;
     login();
+    
 })})
